Create jquery.globalTimeout.js
this was made for when a website has many setTimeouts on it, now it can all be consolidated by first calling $.globalTimeout() on a page, you can set the timeout rate as well as its maximum amount of calls and the event name. Additionally you can clear the timeout to prevent further timeouts. If for some reason you call the $.globalTimeout() again it will reset the max counter. It will also clear the timeout to prevent multiples from running at the same time. Check for the timeout with the event

$(window).on('timeout', function(){
//any function here
});