Create jquery.tarandfeather.js
Untested code to check a string to see if it was matched one previously on a site by the same user. The goal is to prevent a repeat / copy paste post. The point of the plugin is to call $.tarAndFeatherString(postString); every time before posting postString to the server. A random length is used by default to tar and feather making it more difficult for users to skip this test, as well as a random selection of text from the whole string. As with anything client side, if it is your goal to really prevent repeat posts you should also check the post to the server as all client side code can be worked around. This plugin relies on local storage in case the user leaves the page or closes the tab and comes back with the same copy pasta.